server/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pillow
python-pptx
google-generativeai
google-genai
cloudinary
requests
pydantic
aiohttp
